+++ quint/src/runtime/impl/evaluator.ts
@@ -14,7 +14,7 @@
  * @module
  */
 
-import { Either, left, right } from '@sweet-monads/either'
+import { Either, left, mergeInMany, right } from '@sweet-monads/either'
 import { QuintApp, QuintEx } from '../../ir/quintIr'
 import { LookupDefinition, LookupTable } from '../../names/base'
 import { QuintError } from '../../quintError'
@@ -28,7 +28,8 @@ import { zerog } from '../../idGenerator'
 import { List } from 'immutable'
 import { Builder, buildDef, buildExpr, nameWithNamespaces } from './builder'
 import { Presets, SingleBar } from 'cli-progress'
-import { SimulationResult } from '../../simulation'
+import { Outcome, SimulationTrace, getTraceStatistics } from '../../simulation'
+import assert from 'assert'
 
 /**
  * An evaluator for Quint in Node TS runtime.
@@ -135,8 +136,8 @@ export class Evaluator {
    * @param nruns - The number of simulation runs to perform.
    * @param nsteps - The number of steps to perform in each simulation run.
    * @param ntraces - The number of error traces to collect before stopping the simulation.
-   * @returns a boolean expression indicating whether all simulations passed without errors,
-              or an error if the simulation cannot be completed.
+   * @param onTrace - A callback function to be called with trace information for each simulation run.
+   * @returns a simulation outcome with all data to report
    */
   simulate(
     init: QuintEx,
@@ -145,8 +146,9 @@ export class Evaluator {
     witnesses: QuintEx[],
     nruns: number,
     nsteps: number,
-    ntraces: number
-  ): Either<QuintError, SimulationResult> {
+    ntraces: number,
+    onTrace?: (index: number, status: string, vars: string[], states: QuintEx[]) => void
+  ): Outcome {
     let errorsFound = 0
     let failure: QuintError | undefined = undefined
     const startTime = Date.now()
@@ -166,6 +168,7 @@ export class Evaluator {
     const invEval = buildExpr(this.builder, inv)
     const witnessesEvals = witnesses.map(w => buildExpr(this.builder, w))
     const witnessingTraces = new Array(witnesses.length).fill(0)
+    const traceLengths: number[] = []
 
     let runNo = 0
     for (; errorsFound < ntraces && !failure && runNo < nruns; runNo++) {
@@ -203,6 +206,8 @@ export class Evaluator {
 
             const stepResult = stepEval(this.ctx).mapLeft(error => (failure = error))
             if (!isTrue(stepResult)) {
+              traceLengths.push(this.trace.get().length)
+
               // The run cannot be extended. In some cases, this may indicate a deadlock.
               // Since we are doing random simulation, it is very likely
               // that we have not generated good values for extending
@@ -236,6 +241,8 @@ export class Evaluator {
               witnessingTraces[i] = witnessingTraces[i] + 1
             }
           })
+
+          traceLengths.push(this.trace.get().length)
         }
       }
 
@@ -244,9 +251,38 @@ export class Evaluator {
     }
     progressBar.stop()
 
-    return failure
-      ? left(failure)
-      : right({ result: { id: 0n, kind: 'bool', value: errorsFound == 0 }, witnessingTraces, samples: runNo })
+    const results: Either<QuintError[], SimulationTrace[]> = mergeInMany(
+      this.recorder.bestTraces.map((trace, index) => {
+        const maybeEvalResult = trace.frame.result
+        if (maybeEvalResult.isLeft()) {
+          return left(maybeEvalResult.value)
+        }
+        const quintExResult = maybeEvalResult.value.toQuintEx(zerog)
+        assert(quintExResult.kind === 'bool', 'invalid simulation produced non-boolean value ')
+        const simulationSucceeded = quintExResult.value
+        const status = simulationSucceeded ? 'ok' : 'violation'
+        const states = trace.frame.args.map(e => e.toQuintEx(zerog))
+
+        if (onTrace !== undefined) {
+          onTrace(index, status, this.varNames(), states)
+        }
+
+        return right({ states, result: simulationSucceeded, seed: trace.seed })
+      })
+    )
+
+    const runtimeErrors = results.isLeft() ? results.value : []
+
+    let traces = results.isRight() ? results.value : []
+
+    return {
+      status: failure ? 'error' : errorsFound == 0 ? 'ok' : 'violation',
+      errors: failure ? [failure, ...runtimeErrors] : runtimeErrors,
+      bestTraces: traces,
+      witnessingTraces,
+      samples: runNo,
+      traceStatistics: getTraceStatistics(traceLengths),
+    }
   }
 
   /**
@@ -407,12 +443,19 @@ export class Evaluator {
    * @returns the result of the simulation, or an error if the simulation cannot be completed.
    */
   private evaluateSimulation(expr: QuintApp): Either<QuintError, QuintEx> {
+    let result: Outcome
     if (expr.opcode === 'q::testOnce') {
       const [nsteps, ntraces, init, step, inv] = expr.args
-      return this.simulate(init, step, inv, [], 1, toNumber(nsteps), toNumber(ntraces)).map(r => r.result)
+      result = this.simulate(init, step, inv, [], 1, toNumber(nsteps), toNumber(ntraces))
     } else {
       const [nruns, nsteps, ntraces, init, step, inv] = expr.args
-      return this.simulate(init, step, inv, [], toNumber(nruns), toNumber(nsteps), toNumber(ntraces)).map(r => r.result)
+      result = this.simulate(init, step, inv, [], toNumber(nruns), toNumber(nsteps), toNumber(ntraces))
+    }
+
+    if (result.status === 'error') {
+      return left(result.errors[0])
+    } else {
+      return right({ kind: 'str', value: result.status, id: 0n })
     }
   }
 }