+++ quint/src/simulation.ts
@@ -45,6 +45,7 @@ export interface Outcome {
   bestTraces: SimulationTrace[]
   witnessingTraces: number[]
   samples: number
+  traceStatistics: TraceStatistics
 }
 
 /**
@@ -54,4 +55,25 @@ export interface SimulationResult {
   result: QuintEx
   witnessingTraces: number[]
   samples: number
+  traceStatistics: TraceStatistics
+}
+
+export interface TraceStatistics {
+  averageTraceLength: number
+  minTraceLength: number
+  maxTraceLength: number
+}
+
+export function getTraceStatistics(traceLengths: number[]): TraceStatistics {
+  return {
+    maxTraceLength: Math.max(...traceLengths),
+    minTraceLength: Math.min(...traceLengths),
+    averageTraceLength: traceLengths.reduce((a, b) => a + b, 0) / traceLengths.length,
+  }
+}
+
+export function showTraceStatistics(stats: TraceStatistics): string {
+  return `Trace length statistics: max=${stats.maxTraceLength}, min=${
+    stats.minTraceLength
+  }, average=${stats.averageTraceLength.toFixed(2)}`
 }